requirements-test.txt
black==20.8b1
flake8==3.8.*
isort==5.5.*
mypy==0.782
pytest==6.0.*
pytest-cov==2.10.1
pytest-django==3.10.*